emote parse cleanup
